requirements.txt
bintrees==2.0.7
coinbase==2.1.0
fdfgen==0.16.1
gdax==1.0.6
geminipy==0.0.3
pycryptodome==3.6.0
python-dateutil==2.7.2
requests==2.13.0
six==1.10.0
websocket-client==0.40.0
